JSIEVE-92 Code analisys improvements
git-svn-id: https://svn.apache.org/repos/asf/james/jsieve/trunk@1425888 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/jsieve/commands/Require.java
@@ -136,7 +136,7 @@ public class Require extends AbstractPrologCommand {
         List<Argument> args = arguments.getArgumentList();
         if (args.size() != 1)
             throw context.getCoordinate().syntaxException(
-                    new StringBuilder("Exactly 1 argument permitted. Found ").append((int)args.size()));
+                    new StringBuilder("Exactly 1 argument permitted. Found ").append(args.size()));
 
         Argument argument = args.get(0);
         if (!(argument instanceof StringListArgument))
